KeepKey: Update plugin to support WebUSB
