+++ contrib/requirements/requirements-hw.txt
@@ -1,5 +1,5 @@
 Cython>=0.27
 trezor[hidapi]>=0.11.0
-keepkey>=6.0
+keepkey>=6.1
 btchip-python
 hidapi
