+++ plugins/hw_wallet/plugin.py
@@ -27,6 +27,9 @@
 from electroncash.plugins import BasePlugin, hook
 from electroncash.i18n import _
 from electroncash import Transaction
+from electroncash.bitcoin import TYPE_SCRIPT
+from electroncash.util import bfh
+from electroncash.address import OpCodes, Address, Script
 
 class HW_PluginBase(BasePlugin):
     # Derived classes provide:
@@ -52,6 +55,19 @@ class HW_PluginBase(BasePlugin):
             if isinstance(keystore, self.keystore_class):
                 self.device_manager().unpair_xpub(keystore.xpub)
 
+    def show_address(self, wallet, address, keystore=None):
+        pass  # implemented in child classes
+
+    def show_address_helper(self, wallet, address, keystore=None):
+        if keystore is None:
+            keystore = wallet.get_keystore()
+        if not wallet.is_mine(address):
+            keystore.handler.show_error(_('Address not in wallet.'))
+            return False
+        if type(keystore) != self.keystore_class:
+            return False
+        return True
+
 def is_any_tx_output_on_change_branch(tx: Transaction) -> bool:
     if not tx.output_info:
         return False
@@ -61,3 +77,33 @@ def is_any_tx_output_on_change_branch(tx: Transaction) -> bool:
             if info[0][0] == 1:
                 return True
     return False
+
+def validate_op_return_output_and_get_data(output, max_size=220) -> bytes:
+    _type, address, _amount = output
+
+    if _type != TYPE_SCRIPT:
+        raise Exception("Unexpected output type: {}".format(_type))
+
+    ops = Script.get_ops(address.script)
+
+    if not ops[0] == OpCodes.OP_RETURN:
+        raise RuntimeError(_("Only OP_RETURN scripts are supported."))
+
+    if len(ops) > 2 or ops[1][0] not in [OpCodes.OP_PUSHDATA1, OpCodes.OP_PUSHDATA2, OpCodes.OP_PUSHDATA4]:
+        raise RuntimeError(_("OP_RETURN is limited to a single constant push."))
+
+    if len(ops[1][1]) > max_size:
+        raise RuntimeError(_("OP_RETURN data size exceeds the maximum of {} bytes.".format(max_size)))
+
+    if _amount != 0:
+        raise RuntimeError(_("Amount for OP_RETURN output must be zero."))
+
+    return ops[1][1]
+
+def only_hook_if_libraries_available(func):
+    # note: this decorator must wrap @hook, not the other way around,
+    # as 'hook' uses the name of the function it wraps
+    def wrapper(self, *args, **kwargs):
+        if not self.libraries_available: return None
+        return func(self, *args, **kwargs)
+    return wrapper
